experimenting with tests/subtyping
