+++ src/redux/modules/notes.js
@@ -1,5 +1,5 @@
 //@flow
-import type {Note, Store} from "../../models";
+import type {Note} from "../../models";
 
 const ADD_NOTE: 'notes/ADD_NOTE' = 'notes/ADD_NOTE';
 const REMOVE_NOTE: 'notes/REMOVE_NOTE' = 'notes/REMOVE_NOTE';
@@ -21,21 +21,19 @@ const remove = (noteId: number): RemoveNoteAction => {
   return { type: REMOVE_NOTE, noteId }
 };
 
-export const initialStore: Store = {
-  notesList: [
-    { title: "Note #1", description: "Desc #1" },
-    { title: "Note #2", description: "Desc #2" }
-  ]
-};
-
 //*******************************************************************************
 
-const reducer = (state: Store = initialStore, action: NoteAction): Store => {
+// get only needed interface from global redux store
+type NoteStore = {
+  +notesList: Array<Note>
+}
+
+const reducer = (state: NoteStore = initialStore, action: NoteAction): NoteStore => {
   switch (action.type) {
     case ADD_NOTE:
       return { ...state, notesList: [...state.notesList, action.note] };
     case REMOVE_NOTE:
-      return state;
+      return {...state, ...state};
     default:
       return state;
   }
